~9.2

+++ src/controllers/videoController.js
@@ -1,5 +1,6 @@
 import Video from "../models/Video";
-
+import User from "../models/User";
+import res from "express/lib/response";
 export const home = async (req, res) => {
   try {
     const videos = await Video.find({}).sort({ createdAt: "desc" });
@@ -11,7 +12,8 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const id = req.params.id;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
+  console.log(video);
   if (!video) {
     return res.status(400).render("404", { pageTitle: "Video Not Found." });
   }
@@ -20,19 +22,33 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const id = req.params.id;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
-
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  console.log(String(video.owner), _id);
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect("/");
+  }
   res.render("edit", { pageTitle: `Editing `, video });
 };
 
 export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.exists({ _id: id }); // boolean value
   if (!video) {
     return res.render("404", { pageTitle: "Video Not Found." });
   }
-
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -47,14 +63,23 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const {
+    user: { _id },
+  } = req.session;
+  const { path } = req.file;
   const { title, description, hashtags } = req.body;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
+      owner: _id,
       createdAt: Date.now(),
+      fileUrl: path,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.render("upload", {
@@ -66,9 +91,21 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
 
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  }
+  if (String(video.owner) !== _id) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
-
+  user.videos = user.videos.filter((video) => String(video) !== id);
+  user.save();
   res.redirect("/");
 };
